feat: EasyForm type에 RegExp 받을 수 있게 수정
