+++ src/components/CardForm/CardForm.jsx
@@ -1,8 +1,9 @@
 import React from 'react';
 import styled from 'styled-components';
-import EasyForm from './common/EasyForm';
-import CardPreview from './common/CardPreview';
-import ToolTip from './common/ToolTip';
+import EasyForm from '../common/EasyForm';
+import CardPreview from '../common/CardPreview';
+import ToolTip from '../common/ToolTip';
+import cardFormSchema from './cardFormSchema';
 
 const StyledCardForm = styled.form`
   margin: 0;
@@ -113,113 +114,6 @@ const StyledCardFieldContainer = styled.div`
     background-color: #ffc6c6;
   }
 `;
-
-const cardFormSchema = {
-  firstCardNumber: {
-    alias: '카드 번호',
-    type: 'number',
-    initialValue: '',
-    minLength: 4,
-    maxLength: 4,
-    required: true,
-  },
-  secondCardNumber: {
-    alias: '카드 번호',
-    type: 'number',
-    initialValue: '',
-    minLength: 4,
-    maxLength: 4,
-    required: true,
-  },
-  thirdCardNumber: {
-    alias: '카드 번호',
-    type: 'number',
-    initialValue: '',
-    minLength: 4,
-    maxLength: 4,
-    required: true,
-  },
-  fourthCardNumber: {
-    alias: '카드 번호',
-    type: 'number',
-    initialValue: '',
-    minLength: 4,
-    maxLength: 4,
-    required: true,
-  },
-  expiredMonth: {
-    alias: '만료일',
-    type: 'number',
-    initialValue: '',
-    minLength: 2,
-    maxLength: 2,
-    required: true,
-    validation: [
-      {
-        assert: (month) => month >= 1 && month <= 12,
-        errorMessage: '유효하지 않은 월입니다.',
-      },
-    ],
-  },
-  expiredYear: {
-    alias: '만료일',
-    type: 'number',
-    initialValue: '',
-    minLength: 2,
-    maxLength: 2,
-    required: true,
-    validation: [
-      {
-        assert: (year) => year >= new Date().getFullYear() % 100,
-        errorMessage: '유효하지 않은 연도입니다.',
-      },
-    ],
-  },
-  owner: {
-    alias: '카드 소유자 이름',
-    type: 'text',
-    initialValue: '',
-    maxLength: 30,
-  },
-  cvc: {
-    alias: '보안코드 (cvc/cvv)',
-    type: 'number',
-    initialValue: '',
-    minLength: 3,
-    maxLength: 4,
-    required: true,
-  },
-  firstPasswordDigit: {
-    alias: '비밀번호',
-    type: 'number',
-    initialValue: '',
-    minLength: 1,
-    maxLength: 1,
-    required: true,
-  },
-  secondPasswordDigit: {
-    alias: '비밀번호',
-    type: 'number',
-    initialValue: '',
-    minLength: 1,
-    maxLength: 1,
-    required: true,
-  },
-  thirdPasswordDigit: {
-    alias: '비밀번호',
-    type: 'number',
-    initialValue: '*',
-    minLength: 1,
-    maxLength: 1,
-  },
-  fourthPasswordDigit: {
-    alias: '비밀번호',
-    type: 'number',
-    initialValue: '*',
-    minLength: 1,
-    maxLength: 1,
-  },
-};
 
 const getCardInfo = ({
   firstCardNumber,
